Dashboard's Header component done.
we have used shadcn/ui as component library for this application. It is
a lightweight alternative of ui libraries and supports tailwind css. we
also created our custom hook 'useUser.ts' in the hooks folder for using
supabase client.

shadcn also added a components/ui directory in the folder structure.

to work with shadcn we downgraded the application's react version from
react 19 to react 18 along with react-dom.