Add website route completed.
Done with the /add route which is next route after /dashboard. Here we
have 2 steps.In the first step we add the website name and apply some
checks like is this website already added or wrong name etc.
after that we jump to next step where we don't have to do anything
instead here we get a script for our app on which we want to apply the
analytics. After this we will jump on the analytics dashboard of the
specific websiteAdd website route completed.

Done with the /add route which is next route after /dashboard. Here we
have 2 steps.In the first step we add the website name and apply some
checks like is this website already added or wrong name etc.
after that we jump to next step where we don't have to do anything
instead here we get a script for our app on which we want to apply the
analytics. After this we will jump on the analytics dashboard of the
specific website.