Minor updates in different files
Updated the next.config.ts file, changed the images.domain object with
the new way to doing it.
Changed the location of hooks folder outside the app router.

create an api route for calling the script and handling the responses
and various events.
Updated the attributes of the Image component of nextjs for the animated
image (unoptimized = true )